requirements.txt.txt
gradio
chromadb 
langchain_community 
langchain_text_splitters 
langchain_groq 
langchain_chroma 
gradio 
python-dotenv 
pypdf
sentence-transformers